Removing dependency of a proofsTex dir
The sellf binary now outputs the tex files to the
same directory it is running from, and does not
throw an error if a proofsTex dir is not found.

This commit also simplifies a little the file
handling by passing the file name as an argument
to the functions instead of relying on global
variables.